+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * db模块加强异常提示信息
 * 升级可选依赖版本
 * Setting增加getandRemove和getAndRemoveStr方法，并去除db模块中配置键别名获取相关冗余方法
+* FileUtil增加subPath、getPathEle方法（针对JDK7的Path对象封装）
 
 ### Bug修复
 * db模块修复Oracle中传入java.util.Date对象无法识别类型问题（默认按照Timestamp处理）（感谢@【杭州】wiley）